Refactor: Extract `MoviePosterCard` and `MediaItemUi` to a common package
This commit moves `MoviePosterCard` and its data class `MediaItemUi` to a new package `com.moscow.cineverse.screen.component.movie_poster_card` for better organization and reusability across different screens.

**Key changes:**
- Created new package `com.moscow.cineverse.screen.component.movie_poster_card`.
- Moved `MoviePosterCard.kt` to the new package.
- Created `MediaItemUi.kt` in the new package and moved the `MediaItemUi` data class definition there from `ExploreScreenState.kt`.
- Updated import statements in `ShowAllActorMoviesScreen.kt` and `ExploreScreen.kt` to reflect the new location of `MoviePosterCard` and `MediaItemUi`.
- Renamed `moviesUI` to `movies` in `ShowAllActorMoviesState.kt` and `ShowAllActorMoviesScreen.kt` to align with the `MediaItemUi` naming.
- Removed vertical padding from `MovieAppBar` for a more consistent look.
