+++ presentation/src/main/java/com/moscow/cineverse/screen/explore/ExploreScreenState.kt
@@ -2,6 +2,7 @@ package com.moscow.cineverse.screen.explore
 
 import com.moscow.cineverse.designSystem.component.ViewMode
 import com.moscow.cineverse.designSystem.component.tabs.ExploreTabsPages
+import com.moscow.cineverse.screen.component.movie_poster_card.MediaItemUi
 
 data class ExploreScreenState(
 
@@ -46,15 +47,7 @@ data class ExploreScreenState(
         }
     }
 
-    data class MediaItemUi(
-        val id: Int,
-        val title: String,
-        val posterPath: String,
-        val rating: Float,
-        val genres: List<String>,
-        val releaseDate: String,
-        val duration: String
-    )
+
 
     data class ActorUi(
         val title: String,
